Added Composer with autoload
