+++ config/application.config.example.php
@@ -1,11 +1,7 @@
 <?php
 
-$config = [
-    "database" : [
-        "dbname" : "-test.demo-",
-        "host" : "127.0.0.1",
-        "charset" : "utf8mb4",
-        "user" : "-test-",
-        "pass" : "-test-"
+$app_config = [
+    "application" : [
+        
     ]
 ];
