+++ config/encryption.config.example.php
@@ -1,11 +1,7 @@
 <?php
 
-$config = [
-    "database" : [
-        "dbname" : "-test.demo-",
-        "host" : "127.0.0.1",
-        "charset" : "utf8mb4",
-        "user" : "-test-",
-        "pass" : "-test-"
+$enc_config = [
+    "encryption" : [
+        "chars_string" => "HhAzmcl9Dbd1nOeTwYyL47WQvUiZapJs3V5SgENCXtGfMFrBxuPkoRKq60j2I8"
     ]
 ];
