--- src/Application.php
@@ -1,3 +1,0 @@
-<?php
-
-echo "<br />APP<br />"
