+++ src/Core/Application.php
@@ -1,0 +1,14 @@
+<?php
+declare(strict_types=1);
+
+namespace SUS\Core;
+
+class Application {
+    public function __construct() {
+
+    }
+
+    public function run() {
+        echo "App->run()";
+    }
+}
