+++ src/Core/Router.php
@@ -1,0 +1,10 @@
+<?php
+declare(strict_types=1);
+
+namespace SUS\Core;
+
+class Router {
+    public function __construct() {
+
+    }
+}
